virtual_class/requirements.txt
# =============================================================================
# Core Dependencies
# =============================================================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic[email]>=2.8.0

# =============================================================================
# Database & ORM
# =============================================================================
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic==1.13.1

# =============================================================================
# Authentication & Security
# =============================================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# =============================================================================
# LiveKit & Voice Pipeline
# =============================================================================
livekit>=0.17.0
livekit-agents>=1.3.0
livekit-plugins-openai>=0.10.0
livekit-plugins-silero>=0.6.0
websockets>=12.0

livekit-plugins-deepgram==0.6.0

# =============================================================================
# AI & LLM
# =============================================================================
openai>=1.30.0
langchain>=0.1.5
langchain-openai>=0.0.5
langgraph>=0.1.0

# =============================================================================
# Utilities
# =============================================================================
httpx==0.26.0
aiofiles>=24.1.0
